chore: polish docs and tournament flows (#97)
* fix: share session flash messages as Inertia props

Pages already read flash.success/flash.error from page props
(e.g. tournaments/show.tsx), but the values were never exposed
by the Inertia middleware, so the toasts silently never fired
on redirect-with-flash flows.


* feat: block tournament edits once registration has closed

Hide the edit action and guard the edit route so organizers
can't modify a tournament after it leaves draft/registration_open.
Direct URL access redirects to the show page with a flash error,
and service-layer errors surfaced via ValidationException now
render inline on the edit form.


* feat: add league and group-stage tournament formats

Tournaments can now be run as a single-elimination bracket (existing),
a round-robin league where the points table decides the champion, or a
group stage that auto-transitions to a knockout bracket once all group
matches are complete. Reusable standings engine (W/D/L/GF/GA/GD/Pts with
tiebreakers + seeded RNG for unbreakable ties) underpins both new formats.

Schema additions are nullable/defaulted; existing single-elim tournaments
are unaffected (regression test pins behavior).


* feat: improve tournaments index

* chore: commit remaining ui updates

* chore: polish docs and tournament flows

---------
